gestion des droits et sécurité

+++ src/Controller/backOffice/ObservationController.php
@@ -5,6 +5,8 @@ namespace App\Controller\backOffice;
 use App\Entity\Observation;
 use App\Form\ObservationType;
 use App\Repository\ObservationRepository;
+use App\Services\FlashesService;
+use App\Services\UserService;
 use Knp\Component\Pager\PaginatorInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\Controller;
 use Symfony\Component\HttpFoundation\Request;
@@ -17,10 +19,30 @@ use Symfony\Component\Routing\Annotation\Route;
 class ObservationController extends Controller
 {
     /**
+     * @var UserService
+     */
+    private $userService;
+    /**
+     * @var FlashesService
+     */
+    private $flashesService;
+
+    public function __construct(UserService $userService, FlashesService $flashesService)
+    {
+        $this->userService = $userService;
+        $this->flashesService = $flashesService;
+    }
+
+    /**
      * @Route("/", name="observation_index", methods="GET")
      */
-    public function index(ObservationRepository $observationRepository, PaginatorInterface $paginator, Request $request): Response
+    public function observation_index(ObservationRepository $observationRepository, PaginatorInterface $paginator, Request $request): Response
     {
+        if (!$this->userService->isAuthorized($request, __FUNCTION__)){
+            $this->flashesService->setFlashes($this->userService->getFlash());
+            return $this->redirectToRoute('home');
+        };
+
         $q = $request->query->get('q');
         $queryBuilder = $observationRepository->getSpeciesObsWithSearchQueryBuilder($q);
 
@@ -38,8 +60,13 @@ class ObservationController extends Controller
     /**
      * @Route("/new", name="observation_new", methods="GET|POST")
      */
-    public function new(Request $request): Response
+    public function observation_new(Request $request): Response
     {
+        if (!$this->userService->isAuthorized($request, __FUNCTION__)){
+            $this->flashesService->setFlashes($this->userService->getFlash());
+            return $this->redirectToRoute('home');
+        };
+
         $observation = new Observation();
         $form = $this->createForm(ObservationType::class, $observation);
         $form->handleRequest($request);
@@ -61,16 +88,26 @@ class ObservationController extends Controller
     /**
      * @Route("/{id}", name="observation_show", methods="GET")
      */
-    public function show(Observation $observation): Response
+    public function observation_show(Observation $observation, Request $request): Response
     {
+        if (!$this->userService->isAuthorized($request, __FUNCTION__)){
+            $this->flashesService->setFlashes($this->userService->getFlash());
+            return $this->redirectToRoute('home');
+        };
+
         return $this->render('back-office/observation/show.html.twig', ['observation' => $observation]);
     }
 
     /**
      * @Route("/{id}/edit", name="observation_edit", methods="GET|POST")
      */
-    public function edit(Request $request, Observation $observation): Response
+    public function observation_edit(Request $request, Observation $observation): Response
     {
+        if (!$this->userService->isAuthorized($request, __FUNCTION__)){
+            $this->flashesService->setFlashes($this->userService->getFlash());
+            return $this->redirectToRoute('home');
+        };
+
         $form = $this->createForm(ObservationType::class, $observation);
         $form->handleRequest($request);
 
@@ -89,8 +126,13 @@ class ObservationController extends Controller
     /**
      * @Route("/{id}", name="observation_delete", methods="DELETE")
      */
-    public function delete(Request $request, Observation $observation): Response
+    public function observation_delete(Request $request, Observation $observation): Response
     {
+        if (!$this->userService->isAuthorized($request, __FUNCTION__)){
+            $this->flashesService->setFlashes($this->userService->getFlash());
+            return $this->redirectToRoute('home');
+        };
+
         if ($this->isCsrfTokenValid('delete'.$observation->getId(), $request->request->get('_token'))) {
             $em = $this->getDoctrine()->getManager();
             $em->remove($observation);
